holoirc: re-scroll chat log to bottom if necessary
The default behavior for RecycleView and other ListView-like views when resized (for example, by the soft keyboard) is to anchor its contents to the top. This results in the bottom lines of text being obscured, even when the original view was scrolled all the way to the bottom.

By listening for changes in the RecycleView's scroll state, we can detect whether or not the user has scrolled the log to the bottom, and can then re-scroll when detecting a change in the height of the ListView by listening a layout change event.