holoirc: work around weird RecyclerView behavior causing TextViews to become unselectable.
Occasionally, message TextViews would, despite being marked selectable in the view XML, fail to allow selection (a warning gets printed to logcat when this occurs). The root cause of this issue seems the RecyclerView ViewHolder lifecycle, a part of which involves updating the text displayed in the message TextView before the view is attached to a window. This causes the TextView to internally cache the fact that its view hierarchy does not support selection handles, and therefore disallow selection.

The fix is to call setText to "update" the text after the TextView has been reattached to a window. This forces a call to android.widget.Editor#prepareCursorControllers, and re-allows selection.

The timestamp TextView is not affected due to an implementation detail in how the layout of the RecyclerView items are computed; since the message is marked as being to the right of the timestamp, the code path to calculate the width of the timestamp TextView calls prepareCursorControllers for us.